+++ .eslintrc.js
@@ -1,0 +1,22 @@
+module.exports = {
+  root: true,
+  env: {
+    node: true,
+  },
+  parser: "vue-eslint-parser",
+  parserOptions: {
+    parser: "@typescript-eslint/parser",
+    ecmaVersion: 2020,
+    sourceType: "module",
+  },
+  extends: [
+    "plugin:@typescript-eslint/recommended",
+    "eslint:recommended",
+    "prettier/@typescript-eslint",
+    "plugin:prettier/recommended",
+    "plugin:vue/vue3-recommended",
+    "prettier/vue",
+  ],
+  plugins: ["prettier", "@typescript-eslint"],
+  rules: {},
+};