+++ .prettierrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  semi: true,
+  tabWidth: 2,
+  trailingComma: "all"
+};