+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 [0.2.0]
 - TypeScript configuration
+- ESLint / Prettier configuration